Removed implicit conversion: 'int64_t'>'uint64_t'

+++ broadword.hpp
@@ -169,7 +169,7 @@ namespace succinct { namespace broadword {
         if (!x) {
             return false;
         }
-        ret = bit_position(x & -int64_t(x));
+        ret = bit_position(x & -x);
         return true;
 #endif
     }
